face_surveillance_demo/requirements.txt
opencv-python>=4.8.0
flask>=3.0.0
insightface>=0.7.3
onnxruntime>=1.16.0
faiss-cpu>=1.7.4
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
Pillow>=10.0.0
